+++ CHANGES.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `GUIDEMappingContext` now has signals `enabled` and `disabled` which fire when the context is enabled or disabled. This can be useful to trigger some game logic when a context is activated or deactivated. A big thanks goes to [Jonathan Durbin](https://github.com/jonathan-durbin) for providing a PR for this ([#119](https://github.com/godotneers/G.U.I.D.E/pull/119)).
 
+- The virtual cursor modifier now has a setting allowing to synchronize the cursor position with the actual mouse position when activated or deactivated. This is useful when switching between a gamepad and a mouse to avoid jumps in the cursor position. The virtual cursor demo has been updated to make use of this functionality ([#122](https://github.com/godotneers/G.U.I.D.E/issues/122)).
+
 ### Fixed
 - The input detector will now properly detect abort input even when this input was already pressed when starting the detection ([#123](https://github.com/godotneers/G.U.I.D.E/pull/119)).
-
+- When changing mapping contexts, actions will not intermittently report initial values anymore. Instead actions are guaranteed to have the value of their input after the context switch, within the same frame. This avoids issues where actions would briefly report wrong values after a context switch ([#125](https://github.com/godotneers/G.U.I.D.E/issues/125)).
 
 ## [0.8.0] - 2025-09-29
 ### Added